Randomly calling Application.ProcessMessages is pure evil.  Instead, call Form.Repaint or even just Control.Repaint, which does exactly what we are after.
This probably fixes a couple of bugs; most notably that you now shouldn't be able to change the target database of a batch query while the batch is running by clicking in the database tree.  That has happened to me a couple of times over slow connections, even when the query actually looked like it was just about finished.