Include "EXECUTE" as a command for queries which return a resultset. (Executes a prepared statement)
Still thinking of finding a cleaner way to determine whether we got some resultset AFTER having sent a query to the server. But with Zeos we are somewhat limited to detect these sql-commands. If we would work with the mysql-API directly, we could look for mysql_field_count() to see if that's 0 or 1+ and that's it. But I guess we don't live in a perfect world.