Notify user when truncating query.
Decided against wrapping the entire thing in comment markers, since it would ruin coloring and doesn't matter much - there's a ton of other situations which will garble colors in the SQL log.  What we really need is for SynEdit to somehow recognize that a linefeed in the SQL log means that the statement has been terminated and that it should reset it's state.