Store last used database name in registry on FormClose and reselect that in Init.
Works around the issue that you get a "No database selected" after firing a "SELECT * FROM db.table" without having selected a database yet. It's again one of Zeos' automagic queries ("SHOW TABLES LIKE xyz"), which leads to the above error. I thought I can fix that in Zeos by adding the appropriate database-name to the SHOW-statement but that would be quite erroneous as we would have to extract that from the user-written SQL-statement.

The fine thing on this workaround is that it's a nice new small feature as well :)

Thanks to blueyed for the hint.