Bugfix: The export dialog was broken.  Again.  Seems like it is everyone's favorite new hobby ;-).
As long as the new threading code can't handle / queue a bunch of queries from somewhere, it's no good here.  Revert to the way it was before r298.

The hailstorm of SQL queries generated by a remote window during an export does seem like an excellent test case for when the new threading code goes multi-threading / multi-tabbing, though.